Support multiple content directories
This is an initial commit to fix implement this feature, it is still
missing tests. I've tested it on an example site, but I still need to
write some tests. I'm pushing it now just to ask for reviews.

In the
[forum](https://discourse.gohugo.io/t/multiple-contentdirs/7462/4) @bep
recommends to use `UnionFile`, but this solution is not using it, I am
asking for reviews for this approach, but will happily implement it
using `UnionFile` if it's better.

Fixes #3757